feat: unify response
