+++ backend/src/app.controller.ts
@@ -6,9 +6,9 @@ import {
   UnprocessableEntityException,
 } from '@nestjs/common';
 import { AppService } from './app.service';
-import { CarbonEmissionResponseDto } from './carbon-emission-response.dto';
+import { CarbonEmissionRequestDto } from 'src/dto/carbon-emission-request.dto';
+import { CarbonEmissionResponseDto } from 'src/dto/carbon-emission-response.dto';
 import { ApiCreatedResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
-import { CarbonEmissionRequestDto } from './carbon-emission-request.dto';
 
 @Controller('api')
 @ApiTags('탄소 배출량 API')
@@ -22,7 +22,7 @@ export class AppController {
     @Body() body: CarbonEmissionRequestDto,
   ): Promise<CarbonEmissionResponseDto> {
     try {
-      await this.appService.calculateEmission(body.code);
+      return await this.appService.calculateEmission(body.code);
     } catch (error) {
       if (error instanceof UnprocessableEntityException) {
         throw new UnprocessableEntityException(error.message);
@@ -30,7 +30,5 @@ export class AppController {
         throw new InternalServerErrorException();
       }
     }
-
-    return new CarbonEmissionResponseDto(1000, 30000, 120, 0.2, 15, 26000);
   }
 }
