+++ backend/src/app.service.ts
@@ -3,7 +3,6 @@ import { JavaCompilerService } from './java-compiler/java-compiler.service';
 import { JavaRunnerService } from './java-runner/java-runner.service';
 import { CodeService } from './code/code.service';
 import { ConverterService } from './converter/converter.service';
-import { CarbonEmissionsConvertedValuesDto } from './converter/carbon-emissions-converted-values.dto';
 
 @Injectable()
 export class AppService {
@@ -14,9 +13,7 @@ export class AppService {
     private readonly converterService: ConverterService,
   ) {}
 
-  async calculateEmission(
-    input: string,
-  ): Promise<CarbonEmissionsConvertedValuesDto> {
+  async calculateEmission(input: string) {
     const code = await this.codeService.saveCode(input);
     // FIXME: compile result 넘겨주도록 수정
     await this.javaCompilerService.compile(code);
@@ -25,6 +22,6 @@ export class AppService {
     const emission = await this.codeService.calculateEmission(executionResult);
     await this.codeService.updateEmission({ id: code.id, emission: emission });
 
-    return this.converterService.getConvertedCarbonEmissionsValues(emission);
+    return this.converterService.convertCarbonEmission(emission);
   }
 }
